first drat of v2 search job
when running on catalan's constant + 22 PCFs, the default testing_precision of 15 almost catches everything but fails on the PCF that equals 720/(450G-299). thankfully raising testing_precision to 16 catches it, yielding all 253 (23 choose 2, or 23*22/2) possible relations of degree 2 order 1 between catalan's constant and its PCFs. from here it's a job of
1) uploading more PCFs to the DB
2) running v2 on the larger DB (possibly with some filters so that it doesn't include all of the "junk" NamedConstants that we have and are so far completely irrelevant)
also need to add a "sequential" mode where the job is run continuously with its parameters changing between runs